requirements.txt
pandasai>=2.0.0
streamlit>=1.30.0
pandas>=2.0.0
seaborn>=0.13.0
matplotlib>=3.7.0
scipy>=1.11.0
statsmodels>=0.14.0
ydata-profiling>=4.18.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.1.0
pydantic-ai>=0.0.14
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
openpyxl>=3.1.0
numpy>=1.24.0
